[리액트] 조건 검색 페이지 리팩토링 함수들을 별도의 파일로 분리했습니다.
