+++ react-home/home/src/apply_announcement/conditionInfo.js
@@ -1,4 +1,4 @@
-export const conditions = {
+export const conditionInfo = {
     wishRegion:
         {
             category: '희망지역', subcategories: [
